Refactor StaticKey to jwtutils package
移動StaticKeyからcryptokeysパッケージへのjwtutils。
これにより、jwtutilsとのセキュリティおよび独立性の一貫性が強化されます。
また、tailnetの依存関係を減らします。
